% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassThresholdMethods.R
\name{bootBlim}
\alias{bootBlim}
\title{Bootstrap the breakpoint of a hockey-stick stock recruitment curve}
\usage{
bootBlim(
  S,
  R,
  nsim = 100,
  FUN = function(x) {
     sqrt(var(x))/mean(x)
 },
  type = 2,
  ...
)
}
\arguments{
\item{S}{vector of spawning stock biomasses}

\item{R}{vector of recruitment values}

\item{FUN}{function to be applied to summarize Blim across nsim simulations. The default is the CV.}

\item{...}{arguments to pass to calcBlim function}
}
\description{
Bootstrap the breakpoint of a hockey-stick stock recruitment curve
}
