% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassThresholdMethods.R
\name{calcBRP}
\alias{calcBRP}
\title{Calculate biomass reference points (BRPs) in ways based on maximum recruitment (Myers et al. 1994)}
\usage{
calcBRP(S, R, perc = 50, shape = "Ricker", by = 1, maxS = 500)
}
\arguments{
\item{S}{vector of spawning stock biomasses}

\item{R}{vector of recruitment values}

\item{perc}{percent of maximum recruitment (based on S-R curve fit)}

\item{shape}{"Ricker" or "BevertonHolt" shape of S-R curve to fit}

\item{by}{the precision needed for a grid search along the S-R curve.}
}
\description{
Calculate biomass reference points (BRPs) in ways based on maximum recruitment (Myers et al. 1994)
}
\references{
\itemize{
\item R. A. Myers, A. A. Rosenberg, P. M. Mace, N. Barrowman, V. R. Restrepo, In search of thresholds for recruitment overfishing, ICES Journal of Marine Science, Volume 51, Issue 2, 1994, Pages 191-205, https://doi.org/10.1006/jmsc.1994.1020
}
}
