% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassThresholdMethods.R
\name{calcPerc}
\alias{calcPerc}
\title{Given parameters a and b to define a curve, and a target value of S, find the percentage of maxR corresponding to that S value.}
\usage{
calcPerc(a, b, Starget, shape = "Ricker", by = NULL, maxS = 10000)
}
\arguments{
\item{shape}{can be "hockey", "Ricker", or "Beverton-Holt"}

\item{by}{the precision needed for a grid search for target. If missing, 100 points from 0 to maxS are tried.}
}
\description{
Given parameters a and b to define a curve, and a target value of S, find the percentage of maxR corresponding to that S value.
}
\details{
\itemize{
\item "hockey" shape must have pars=c(a,b) where (a,b) is the inflection point
\item "contHockey" shape must have pars=c(beta, S*) following the notation of Mesnil & Rochet 2010 (eqn 4)
\item "Ricker shape" must have pars=c(a,b) where expected recruitment is a*S*exp(-b*S)
\item "BevertonHolt" shape must have pars=c(a,b) where expected recruitment is a*S/(1+b*S)
\item tail is a named list containing a probability ("prob") of observing a bonanza year and a multiplier ("mult") to indicate how a bonanza compares to the expected mean. e.g. tail=list(prob=0.2, mult=1.5)
}
}
