% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMethods.R
\name{simR}
\alias{simR}
\title{Simulate recruitment}
\usage{
simR(S, shape = "contHockey", pars = c(10, 100), g = 1, var = 0,
  cor = 0, varlog = NULL, tail = NULL)
}
\arguments{
\item{shape}{can be "hockey", "Ricker", or "Beverton-Holt"}

\item{pars}{vector of parameters for the specified shape (see details)}

\item{g}{is only used for the hyperbolic continuous hockey stick model (see details)}

\item{var}{used to specify variance on the natural scale}

\item{varlog}{used to specify variance on the log scale}

\item{tail}{controls bonanzas (i.e. extra good recruitment events) (see details)}
}
\description{
Simulate recruitment
}
\details{
\itemize{
\item "hockey" shape must have pars=c(a,b) where (a,b) is the inflection point
\item "contHockey" shape must have pars=c(beta, S*) following the notation of Mesnil & Rochet 2010 (eqn 4)
\item "Ricker shape" must have pars=c(a,b) where expected recruitment is a*S*exp(-b*S)
\item "BevertonHolt" shape must have pars=c(a,b) where expected recruitment is a*S/(1+b*S)
\item tail is a named list containing a probability ("prob") of observing a bonanza year and a multiplier ("mult") to indicate how a bonanza compares to the expected mean. e.g. tail=list(prob=0.2, mult=1.5)
}
}
