% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMethods.R
\name{simS}
\alias{simS}
\title{Simulate SSB}
\usage{
simS(n = 20, Sdist = "unif", pars = c(10, 400))
}
\arguments{
\item{Sdist}{the distribution of SSB, can be "unif" or "lognormal"}
}
\description{
Simulate SSB
}
\details{
\itemize{
\item unif uses pars=(minS, maxS)
\item lognormal uses pars=(mean, sd)
}
}
