% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassThresholdMethods.R
\name{calcBlim}
\alias{calcBlim}
\title{Calculate Blim in ways dexcribed by ICES Advice Technical Guidelines Table 12.4.3.1.3}
\usage{
calcBlim(S, R, quant = 0.75, type = 1, g = 1, by = 1)
}
\arguments{
\item{S}{vector of spawning stock biomasses}

\item{R}{vector of recruitment values}

\item{quant}{quantile above which recruitment is considered to be "large" in a "Spasmodic stock" (language from ICES Advice Technical Guidelines)}

\item{type}{way of calculating Blim.
1 is the minimum S that gives good recruitment.
2 is the estimated S for a breakpoint in a hockey stick model.
2.1 is estimated S for inflection point in a hyperbolic hockey-stick model.
5 is the minimum observed S (Blim=Bloss).}

\item{g}{is the assumed smoothing parameter in the hyperbolic hockey-stick model.}

\item{by}{the precision needed for a grid search for breakpoint in a hockey-stick model.}
}
\description{
Calculate Blim in ways dexcribed by ICES Advice Technical Guidelines Table 12.4.3.1.3
}
